refactor: move ai code under core ai directory
