+++ demos/vite-vue3/package.json
@@ -22,7 +22,7 @@
     "@vitejs/plugin-vue-jsx": "^3.0.1",
     "code-inspector-plugin": "workspace:*",
     "install": "^0.13.0",
-    "node-pty": "^1.0.0",
+    "node-pty": "^1.1.0",
     "npm": "^11.11.0",
     "pug": "^3.0.2",
     "typescript": "^4.9.3",
