+++ packages/core/package.json
@@ -57,8 +57,9 @@
     "chalk": "^4.1.1",
     "dotenv": "^16.1.4",
     "launch-ide": "1.4.5",
-    "marked": "^17.0.1",
-    "portfinder": "^1.0.28"
+    "node-pty": "^1.1.0",
+    "portfinder": "^1.0.28",
+    "ws": "^8.18.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
@@ -71,6 +72,7 @@
     "@vue/compiler-sfc": "^3.3.4",
     "lit": "^2.6.1",
     "magic-string": "^0.30.0",
+    "marked": "^17.0.1",
     "rollup-plugin-terser": "^7.0.2",
     "svelte": "^4.2.7",
     "typescript": "^4.9.3",
@@ -81,8 +83,6 @@
   "optionalDependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.2.29",
     "@openai/codex-sdk": "^0.106.0",
-    "@opencode-ai/sdk": "^1.2.24",
-    "node-pty": "^1.0.0",
-    "ws": "^8.18.0"
+    "@opencode-ai/sdk": "^1.2.24"
   }
 }
