+++ packages/core/src/ai/client/ai-terminal.ts
