+++ packages/core/types/ai/client/ai-terminal.d.ts
