+++ packages/core/types/ai/client/ai.d.ts
