+++ packages/core/types/ai/server/ai-provider-codex.d.ts
@@ -1,4 +1,4 @@
-import type { CodexOptions } from '../shared';
+import type { CodexOptions } from '../../shared';
 import type { AIContext, AIMessage } from './ai';
 import type { ProviderCallbacks, ProviderResult } from './ai-provider-claude';
 import { __TEST_ONLY__ } from './ai-provider-common';
