+++ packages/core/types/ai/server/ai-provider-opencode.d.ts
@@ -1,4 +1,4 @@
-import type { OpenCodeOptions } from '../shared';
+import type { OpenCodeOptions } from '../../shared';
 import type { AIContext, AIMessage } from './ai';
 import type { ProviderCallbacks, ProviderResult } from './ai-provider-claude';
 import { type CodexProviderRuntime } from './ai-provider-common';
