+++ packages/core/types/ai/server/ai-terminal.d.ts
