+++ test/core/server/ai/ai-branches.test.ts
@@ -7,17 +7,17 @@ const mockGetCodexModelInfo = vi.hoisted(() => vi.fn(async () => ''));
 const mockGetOpenCodeModelInfo = vi.hoisted(() => vi.fn(async () => ''));
 const mockGetClaudeModelInfo = vi.hoisted(() => vi.fn(async () => ''));
 
-vi.mock('@/core/src/server/ai-provider-codex', () => ({
+vi.mock('@/core/src/ai/server/ai-provider-codex', () => ({
   handleCodexRequest: mockHandleCodexRequest,
   getModelInfo: mockGetCodexModelInfo,
 }));
 
-vi.mock('@/core/src/server/ai-provider-opencode', () => ({
+vi.mock('@/core/src/ai/server/ai-provider-opencode', () => ({
   handleOpenCodeRequest: mockHandleOpenCodeRequest,
   getModelInfo: mockGetOpenCodeModelInfo,
 }));
 
-vi.mock('@/core/src/server/ai-provider-claude', () => ({
+vi.mock('@/core/src/ai/server/ai-provider-claude', () => ({
   handleClaudeRequest: mockHandleClaudeRequest,
   getModelInfo: mockGetClaudeModelInfo,
 }));
@@ -28,7 +28,7 @@ import {
   getAIOptions,
   handleAIRuntimeAbortRequest,
   handleAIRuntimeStreamRequest,
-} from '@/core/src/server/ai';
+} from '@/core/src/ai/server/ai';
 
 function createMockReq(rawBody: string) {
   const listeners: Record<string, Function[]> = {};
