+++ vitest.config.ts
@@ -11,14 +11,14 @@ export default defineConfig({
       exclude: [
         '**/*.d.ts',
         'packages/core/src/shared/type.ts',
-        'packages/core/src/client/ai-terminal.ts',
-        'packages/core/src/client/ai.ts',
+        'packages/core/src/ai/client/ai-terminal.ts',
+        'packages/core/src/ai/client/ai.ts',
         'packages/core/src/client/i18n.ts',
         'packages/core/src/client/index.ts',
-        'packages/core/src/server/ai-provider-common.ts',
-        'packages/core/src/server/ai-provider-opencode.ts',
-        'packages/core/src/server/ai-terminal.ts',
-        'packages/core/src/server/ai.ts',
+        'packages/core/src/ai/server/ai-provider-common.ts',
+        'packages/core/src/ai/server/ai-provider-opencode.ts',
+        'packages/core/src/ai/server/ai-terminal.ts',
+        'packages/core/src/ai/server/ai.ts',
       ],
     },
     alias: {
